+++ changelog.md
@@ -12,6 +12,10 @@
   * FdFitter: `FdFit.plot()` now raises a `KeyError` when trying to plot data that does not exist.
   * FdFitter: `FdFit.plot()` now raises a `RuntimeError` when trying to plot a fit with multiple models without selecting a model using angular brackets `[]` first.
   * FdFitter: `FdFit.profile()` now raises a `ValueError` when `max_step <= min_step` or `max_chi2_step <= min_chi2_step`.
+  * FdFitter: when inverting a model with `interpolate=True`, Pylake now raises a `ValueError` if the minimum or maximum is not finite.
+  * FdFitter: a `ValueError` is raised when adding incompatible models.
+  * FdFitter: When adding data to a fit, adding data with an unequal number of points for the dependent and independent variable will now raise a `ValueError`.
+  * FdFitter: When adding data to a fit, adding data with more than one dimension will raise a `ValueError`. 
   
 #### New features